[message] extract internal API into new _httpmessage.h file
